treewide: Make multiple runqueue task counters 32-bit
Creating 2**32 tasks is impossible due to futex pid limits and wasteful
anyway. Nobody has done it.

If total number of tasks can't exceed 2**32 (and less due to futex pid
limits), then per-runqueue counters can't as well.

This patchset has been sponsored by REX Prefix Eradication Society.

* sched: Make nr_running() return 32-bit value
* sched: Make nr_iowait() return 32-bit value
* sched: Make nr_iowait_cpu() return 32-bit value
* sched: Make multiple runqueue task counters 32-bit

Link: https://lore.kernel.org/r/20210422200228.1423391-1-adobriyan@gmail.com
Link: https://lore.kernel.org/r/20210422200228.1423391-2-adobriyan@gmail.com
Link: https://lore.kernel.org/r/20210422200228.1423391-3-adobriyan@gmail.com
Link: https://lore.kernel.org/r/20210422200228.1423391-4-adobriyan@gmail.com

+++ kernel/sched/core.c
@@ -3733,9 +3733,9 @@ context_switch(struct rq *rq, struct task_struct *prev,
  * externally visible scheduler statistics: current number of runnable
  * threads, total number of context switches performed since bootup.
  */
-unsigned long nr_running(void)
+unsigned int nr_running(void)
 {
-	unsigned long i, sum = 0;
+	unsigned int i, sum = 0;
 
 	for_each_online_cpu(i)
 		sum += cpu_rq(i)->nr_running;
@@ -3815,9 +3815,9 @@ unsigned long nr_iowait_cpu(int cpu)
  * Task CPU affinities can make all that even more 'interesting'.
  */
 
-unsigned long nr_iowait(void)
+unsigned int nr_iowait(void)
 {
-	unsigned long i, sum = 0;
+	unsigned int i, sum = 0;
 
 	for_each_possible_cpu(i)
 		sum += nr_iowait_cpu(i);
